+++ src/__tests__/types-tests.ts
@@ -137,3 +137,53 @@ test("Address filters", async () => {
     await testField("messages", "dst")
     await testField("transactions", "account_addr")
 })
+
+test("Non Exists acc_type for missing accounts", async () => {
+    const id1 = `0:${"1".repeat(64)}`
+    const id2 = `-1:${"2".repeat(64)}`
+    const idEq = (
+        await testServerQuery<Record<string, Record<string, unknown>[]>>(
+            `query { accounts(filter:{id:{eq:"${id1}"}})
+            { id acc_type acc_type_name balance workchain_id } }`,
+        )
+    )["accounts"]
+
+    const accounts = (ids: string[]) =>
+        ids.map(id => ({
+            id,
+            acc_type: 3,
+            acc_type_name: "NonExist",
+            balance: "0x0",
+            workchain_id: Number(id.split(":")[0]),
+        }))
+    expect(idEq).toEqual(accounts([id1]))
+
+    const idIn = (
+        await testServerQuery<Record<string, Record<string, unknown>[]>>(
+            `query { accounts(filter:{id:{in:["${id1}", "${id2}"]}})
+            { id acc_type acc_type_name balance workchain_id } }`,
+        )
+    )["accounts"]
+    expect(idIn).toEqual(accounts([id1, id2]))
+
+    const blockchainAccountAddress = (
+        await testServerQuery<Record<string, Record<string, unknown>[]>>(
+            `query { blockchain { account(address:"${id1}")
+            { info { id acc_type acc_type_name balance workchain_id } transactions { edges { node { hash } } } messages { edges { node { hash } } } } } }`,
+        )
+    )["blockchain"]
+
+    expect(blockchainAccountAddress).toEqual({
+        account: {
+            info: {
+                id: `account/${id1}`,
+                acc_type: 3,
+                acc_type_name: "NonExist",
+                balance: "0x0",
+                workchain_id: Number(id1.split(":")[0]),
+            },
+            transactions: { edges: [] },
+            messages: { edges: [] },
+        },
+    })
+})